Added more features i deemed helpful
Retweets and likes are obviously a huge indicator for how viral a tweet has become. Somehow adding "favorite_count", i assume its the amount of likes, is destroying the search of the best K. The best K is apparently 0 with an accuracy of around 0.94.

A huge gap of accuracy has occured between the trained data and test data, I succesfully reduced it from 0.02 to around 0.003